0/0.1/demo/requirements.txt
browserstack-sdk
Appium-Python-Client==2.6.1;python_version >= '3.0'
